+++ CHANGELOG.md
@@ -3,5 +3,6 @@
 ## Unreleased
 ### Added
  - `Node` class
+ - `Graph` class
  - Merge capabilities
  - `update`, `historical`, `deferred` events.